backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "sound-competition-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    # Database
    "sqlalchemy>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pgvector>=0.2.5",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    # Config
    "pydantic-settings>=2.3.0",
    # Redis / Celery
    "redis>=5.0.6",
    "celery>=5.4.0",
    # Cloudinary
    "cloudinary>=1.40.0",
    # Audio fingerprinting
    "pyacoustid>=1.3.0",
    "librosa>=0.10.2",
    "numpy>=1.26.0",
    # HTTP
    "httpx>=0.27.0",
    "aiofiles>=23.2.1",
    # Rate limiting
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]
